Decouple GDTF geometry types from loader header
